% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{normalize_deed_colnames}
\alias{normalize_deed_colnames}
\title{Normalize DEED column names}
\usage{
normalize_deed_colnames(colnames)
}
\arguments{
\item{colnames}{Character vector of column names}
}
\value{
Normalized column names
}
\description{
Converts DEED Excel column names to standardized format.
}
\keyword{internal}
