% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{normalize_ethnicity_values}
\alias{normalize_ethnicity_values}
\title{Normalize ethnicity values}
\usage{
normalize_ethnicity_values(x)
}
\arguments{
\item{x}{Vector of ethnicity strings}
}
\value{
Normalized ethnicity names
}
\description{
Converts ethnicity string values from DEED to standardized column names.
}
\keyword{internal}
