% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{standardize_old_format_ethnicity}
\alias{standardize_old_format_ethnicity}
\title{Standardize old format ethnicity enrollment data}
\usage{
standardize_old_format_ethnicity(df)
}
\arguments{
\item{df}{Data frame from old format file}
}
\value{
Data frame with standardized column names
}
\description{
Converts 2021-2023 format (ID, District / School, Ethnicity) to 2024+ format
(Type, id, District/School, Ethnicity).
}
\keyword{internal}
